+++ CHANGELOG.md
@@ -19,6 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `partial-recall-{max-age=>intermediate-term}` and
   `partial-recall-{reclaim-min-age=>short-term}`. They can both be set
   to nil to disable their usage. `partial-recall-reclaim` was removed.
+- Hooks `partial-recall-{probe,permanence-change,after-insert}-hook`
+  now use `run-hook-with-args` and pass sensible arguments to their
+  functions (the memory, the moment and permanence and the moment
+  respectively).
 
 ## [v0.8.6]